# tooling/sparta/packages/utils/openapi/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sparta API",
    "version": "1.0.0",
    "description": "…"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Sparta API"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message describing the issue."
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "VerifyResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "verified": {
            "type": "boolean",
            "description": "Whether the overall verification (score + role assignment) was successful."
          },
          "score": {
            "type": "number",
            "description": "The user's Gitcoin Passport score."
          },
          "roleAssigned": {
            "type": "boolean",
            "description": "Whether the Discord role was successfully assigned/updated."
          },
          "address": {
            "type": "string",
            "description": "The wallet address recovered from the signature."
          },
          "status": {
            "type": "string",
            "description": "Final status of the verification process (e.g., verification_complete, verification_failed)."
          },
          "message": {
            "type": "string",
            "description": "A user-friendly message summarizing the result."
          },
          "minimumRequiredScore": {
            "type": "number",
            "description": "The minimum score required for verification."
          }
        },
        "required": [
          "success",
          "verified",
          "score",
          "roleAssigned",
          "address",
          "status",
          "message",
          "minimumRequiredScore"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/human/verify": {
      "post": {
        "summary": "Verify a wallet signature",
        "description": "Verify a wallet signature and process Passport verification",
        "tags": [
          "Human"
        ],
        "operationId": "verifySignature",
        "parameters": [
          {
            "in": "query",
            "name": "verificationId",
            "schema": {
              "type": "string"
            },
            "description": "The verification ID (can also be provided in body)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "signature"
                ],
                "properties": {
                  "signature": {
                    "type": "string",
                    "description": "Wallet signature"
                  },
                  "verificationId": {
                    "type": "string",
                    "description": "The verification ID (if not provided in query)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signature verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing parameters or invalid signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Verification not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/human/status/discord/{discordUserId}": {
      "get": {
        "summary": "Check verification status by Discord user ID",
        "description": "Check the human verification status of a user by their Discord ID",
        "tags": [
          "Human"
        ],
        "operationId": "getStatusByDiscordId",
        "parameters": [
          {
            "in": "path",
            "name": "discordUserId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Discord user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Status returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/human/status/verification/{verificationId}": {
      "get": {
        "summary": "Check verification status by verification ID",
        "description": "Check the human verification status of a user by their verification ID",
        "tags": [
          "Human"
        ],
        "operationId": "getStatusByVerificationId",
        "parameters": [
          {
            "in": "path",
            "name": "verificationId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Verification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Status returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/human/score": {
      "get": {
        "summary": "Get passport score for a given address and verification",
        "description": "Fetches the Gitcoin Passport score for the wallet address associated with a verification ID",
        "tags": [
          "Human"
        ],
        "operationId": "getScore",
        "parameters": [
          {
            "in": "query",
            "name": "verificationId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The verification ID obtained during verification initiation."
          },
          {
            "in": "query",
            "name": "address",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The wallet address to check (case-insensitive comparison)."
          }
        ],
        "responses": {
          "200": {
            "description": "Score retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Verification not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error during score fetching or processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/verification/{verificationId}": {
      "get": {
        "summary": "Get a user by verification ID",
        "description": "Retrieve a user by their Human Passport verification ID",
        "tags": [
          "Users"
        ],
        "operationId": "getUserByVerificationId",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "verificationId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Human Passport verification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Success status"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing verification ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "User not found with this verification ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
...
}